requirements.txt
PyYAML==5.3
tabulate==0.8.6
python-decouple==3.3